+++ assets/javascripts/kitten/components/notifications/alert.test.js
@@ -2,9 +2,16 @@ import React from 'react'
 import classNames from 'classnames'
 import { expect } from 'chai'
 import { shallow, mount } from 'enzyme'
+import sinon from 'sinon'
 import { Alert } from 'kitten/components/notifications/alert'
 
 describe('<Alert />', () => {
+  const sandbox = sinon.sandbox.create()
+
+  afterEach(() => {
+    sandbox.restore()
+  })
+
   describe('by default', () => {
     const alert = shallow(<Alert />)
 
@@ -103,11 +110,50 @@ describe('<Alert />', () => {
     })
   })
 
+  describe('with onClose prop', () => {
+    let onCloseSpy
+    let alertComponent
+
+    before(() => {
+      onCloseSpy = sandbox.spy()
+      alertComponent = mount(
+        <Alert
+          closeButton
+          onClose={ onCloseSpy }
+        />
+      )
+
+      alertComponent.instance().handleAnimationEnd()
+    })
+
+    it('calls onClose prop callback', () => {
+      expect(onCloseSpy.calledOnce).to.equal(true)
+    })
+  })
+
+  describe('simulates onAnimationEnd event', () => {
+    let onCloseSpy
+    let alertComponent
+
+    before(() => {
+      onCloseSpy = sandbox.spy()
+      alertComponent = mount(<Alert onClose={ onCloseSpy } />)
+
+      alertComponent.simulate('animationEnd')
+    })
+
+    it('has handleAnimationEnd in onAnimationEnd attribute', () => {
+      expect(onCloseSpy.calledOnce).to.equal(true)
+    })
+  })
+
   describe('simulates click event on CloseButton', () => {
     const alert = mount(<Alert closeButton />)
     const closeButton = alert.find('CloseButton')
 
-    closeButton.simulate('click')
+    before(() => {
+      closeButton.simulate('click')
+    })
 
     it('changes state', () => {
       expect(alert.state().show).to.be.false