+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Use `k-u-weight-regular` instead of `k-u-strong` in `ProjectCard` component.
+- Feature: Update placeholder `font-weight` on `LocationInput` component.
+- Feature: Update `font-weight` on `LinkBox` component.
 
 ## [14.5.0] - 2017-11-10