+++ CHANGELOG.md
@@ -87,6 +87,10 @@ const resolvingPaths = kittenComponents.jsPaths
                        .concat(kittenComponents.fontsPaths)
 ```
 
+- Fix: Use `k-u-weight-regular` instead of `k-u-strong` in `ProjectCard` component.
+- Feature: Update placeholder `font-weight` on `LocationInput` component.
+- Feature: Update `font-weight` on `LinkBox` component.
+
 ## [14.5.0] - 2017-11-10
 
 Features: