+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Update error `line-height` on `FormInfo` component.
 - Feature: Add `onClose` prop to `Alert` component.
 - Fix: Use `k-u-weight-regular` instead of `k-u-strong` in `ProjectCard` component.
 - Feature: Update placeholder `font-weight` on `LocationInput` component.